Improve StrictMath support (#10186)
Fixes #9754

Adds all missing methods as simple aliases to the respective `Math`
methods, consistent with the rest of the class.

Fixes a discrepancy in the return type of `Math.floorMod(long, int)`.

+++ user/super/com/google/gwt/emul/java/lang/Math.java
@@ -150,8 +150,8 @@ public final class Math {
     return ((dividend % divisor) + divisor) % divisor;
   }
 
-  public static long floorMod(long dividend, int divisor) {
-    return floorMod(dividend, (long) divisor);
+  public static int floorMod(long dividend, int divisor) {
+    return (int) floorMod(dividend, (long) divisor);
   }
 
   @SuppressWarnings("CheckStyle.MethodName")
